# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Brinecrypt API
  description: |
    Brinecrypt is a secure resource management and encryption service designed to provide a centralized and secure way to store and manage sensitive configuration data, secrets, and other resources.

    ### Core Concepts

    *   **Resources:** Named entities stored within a **Namespace**. Resources can be of type `cleartext` (stored as is) or `encrypted` (transparently encrypted/decrypted).
    *   **Namespaces:** Logical groupings for resources. The `_` namespace is reserved for internal system resources like users.
    *   **Versions:** Every time a resource is updated via `PUT`, a new version is created. Brinecrypt maintains a history of all versions.
    *   **Principals:** Entities that can be granted permissions. These include **Users** and **Service Accounts** (SAs).
    *   **Permissions:** Granular access control based on a `ResourcePattern` (e.g., `prod/*`), a `Verb` (`list`, `read`, `write`, `delete`), and an optional expiry time.

    ### Authentication & Authorization

    Brinecrypt supports multiple authentication methods:
    1.  **Session Tokens:** Obtained via `/auth/login`. Used for interactive sessions.
    2.  **Personal Access Tokens (PATs):** Long-lived tokens issued by users for programmatic access.
    3.  **Capability Tokens:** Restricted tokens issued with a specific set of permissions, useful for delegation or narrow-scoped tasks.
...

...

    ### Encryption

...
  version: 1.0.0
servers:
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh session token using a refresh token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshSessionRequest'
      responses:
        '200':
          description: Successfully refreshed tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseBody'
        '400':
          description: Malformed input
        '401':
          description: Unauthorized or expired refresh token
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: List all users
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '403':
          description: Forbidden
        '500':
          description: Internal server error
    post:
      tags:
        - Admin
      summary: Create a new user
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '204':
          description: User successfully created
        '400':
          description: Bad request
        '403':
          description: Forbidden (requires administrative privileges)
        '500':
          description: Internal server error
  /admin/users/{name}:
    get:
      tags:
        - Admin
      summary: Get user by name
      operationId: getUserByName
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          description: Forbidden
        '404':
          description: User not found
    delete:
      tags:
        - Admin
      summary: Delete user by name
      operationId: deleteUserByName
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User successfully deleted
        '403':
          description: Forbidden
        '500':
          description: Internal server error
  /admin/permissions:
    post:
      tags:
        - Admin
      summary: Grant permissions to a principal (user or service account)
      operationId: grantPermissions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionsRequest'
      responses:
        '204':
          description: Permissions successfully granted
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Principal not found
    delete:
      tags:
        - Admin
      summary: Revoke permissions from a principal
      operationId: revokePermissions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionsRequest'
      responses:
        '204':
          description: Permissions successfully revoked
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Principal not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Verb:
      type: string
      enum: [list, read, write, delete]
      description: Allowed operations on resources.
...
    Permission:
      type: object
      properties:
        resource_pattern:
          type: string
          example: "default/*"
        verb:
          $ref: '#/components/schemas/Verb'
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
      required:
        - resource_pattern
        - verb
        - created_at
    PermissionEntry:
      type: object
      properties:
        verb:
          $ref: '#/components/schemas/Verb'
        resource_pattern:
          type: string
          example: "default/my-secret"
        expires_at:
          type: string
          format: date-time
      required:
        - verb
        - resource_pattern
...
    LoginResponseBody:
      type: object
      properties:
        session_token:
          type: string
        refresh_token:
          type: string
      required:
        - session_token
        - refresh_token
    RefreshSessionRequest:
      type: object
      properties:
        token:
          type: string
      required:
        - token
...
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        pass:
          type: string
      required:
        - name
        - pass
    PermissionsRequest:
      type: object
      properties:
        principal:
          type: string
          description: "Principal identifier (e.g., 'user:username' or 'sa:namespace:name')"
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionEntry'
      required:
        - principal
        - permissions
    UserResponse:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
      required:
        - name
        - email
        - created_at
...